Cross platform support + code cleanup (#54)
* Removed a .gitkeep

* Cleaned the swapchain of dead code

* Removed useless comment

* Moved all of the systems into a single dir

* Automatic dep installation for linux

* Added algorithm header for stricter compilation

* Actual error message for the not found shader

* Added cmake in the installs, and adviced to execute cd and the program

* Fixed a race condition issue on the image semaphore

* Increased the page size of the descriptors for ImGui

* Made compilation possible on windows

* Cross platform clamp issue fixed

* Removed auto add component

* Removed all of the implicit casts

* Fixed the last warnings on windows

* Fixed last warnigs with linux

* Fixed the name issue in the EntityHierarchy

* Added names to the entities of the default scene

* Fixed the crash when resizing after having deleted an entity and fixed the preset for linux

* Fixed the message in the setup.sh

* Good names for the labels of the components

* Fixed the two syntax issues

+++ src/api/vulkan/Pipeline.cpp
@@ -5,7 +5,7 @@
 /*  Project: Hel Engine                                                       */
 /*  Created: 2026/01/13 19:39:15 by hle-hena                                  */
 /*                                                                            */
-/*  Last Modified: 2026/04/08 18:21:18                                        */
+/*  Last Modified: 2026/05/01 12:18:01                                        */
 /*             By: hle-hena                                                   */
 /*                                                                            */
 /*    -----                                                                   */
@@ -48,20 +48,20 @@ bool	Pipeline::createGraphicsPipeline(PipelineConfig &config,
 	vertexInputInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_VERTEX_INPUT_STATE_CREATE_INFO;
 	vertexInputInfo.pVertexAttributeDescriptions = config.attributeDescription.data();
 	vertexInputInfo.pVertexBindingDescriptions = config.bindingDescription.data();
-	vertexInputInfo.vertexAttributeDescriptionCount = config.attributeDescription.size();
-	vertexInputInfo.vertexBindingDescriptionCount = config.bindingDescription.size();
+	vertexInputInfo.vertexAttributeDescriptionCount = static_cast<uint32_t>(config.attributeDescription.size());
+	vertexInputInfo.vertexBindingDescriptionCount = static_cast<uint32_t>(config.bindingDescription.size());
 
 	VkPipelineColorBlendStateCreateInfo	blendStateInfo{};
 	blendStateInfo.sType = VK_STRUCTURE_TYPE_PIPELINE_COLOR_BLEND_STATE_CREATE_INFO;
 	blendStateInfo.logicOpEnable = VK_FALSE;
-	blendStateInfo.attachmentCount = config.colorBlendAttachment.size();
+	blendStateInfo.attachmentCount = static_cast<uint32_t>(config.colorBlendAttachment.size());
 	blendStateInfo.pAttachments = config.colorBlendAttachment.data();
 
 	VkGraphicsPipelineCreateInfo	pipelineInfo;
 	pipelineInfo.sType = VK_STRUCTURE_TYPE_GRAPHICS_PIPELINE_CREATE_INFO;
 	pipelineInfo.pNext = &config.renderingInfo;
 	pipelineInfo.flags = 0;
-	pipelineInfo.stageCount = stageInfo.size();
+	pipelineInfo.stageCount = static_cast<uint32_t>(stageInfo.size());
 	pipelineInfo.pStages = stageInfo.data();
 	pipelineInfo.pVertexInputState = &vertexInputInfo;
 	pipelineInfo.pInputAssemblyState = &config.inputAssemblyInfo;
