+++ src/ecs/assets/Geometry.tpp
@@ -25,13 +25,15 @@ std::unique_ptr<Buffer>	Geometry::createBuffer(Device &device,
 											std::vector<T> data,
 											VkBufferUsageFlags usage) {
 	VkDeviceSize	size = sizeof(T) * data.size();
-	auto	stagingBuffer = Buffer::create(device, sizeof(T), data.size(),
+	auto	stagingBuffer = Buffer::create(device, sizeof(T),
+		static_cast<uint32_t>(data.size()),
 		VK_BUFFER_USAGE_TRANSFER_SRC_BIT, VMA_MEMORY_USAGE_AUTO_PREFER_HOST,
 		VMA_ALLOCATION_CREATE_HOST_ACCESS_SEQUENTIAL_WRITE_BIT);
 
 	stagingBuffer->writeToBuffer(data.data(), size);
 
-	auto	buffer = Buffer::create(device, sizeof(T), data.size(),
+	auto	buffer = Buffer::create(device, sizeof(T),
+		static_cast<uint32_t>(data.size()),
 		VK_BUFFER_USAGE_TRANSFER_DST_BIT | usage,
 		VMA_MEMORY_USAGE_AUTO_PREFER_DEVICE);
